requirement.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.5
django-crispy-forms==1.9.0
django-filter==2.3.0
django-filters==0.2.1
django-jsonfield==1.4.0
django-mysql==3.5.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
phonenumbers==8.12.6
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
